Moving more functionality outside of SudokuBoard.tsx
