+++ sudokuru/app/Components/SudokuBoard/Core/Components/NumberControl.tsx
@@ -5,6 +5,7 @@ import React from "react";
 import { getCellSize } from "../Functions/BoardFunctions";
 import { LinearGradient } from "expo-linear-gradient";
 import { PreferencesContext } from "../../../../Contexts/PreferencesContext";
+import { SudokuObjectProps } from "../../../../Functions/LocalDatabase";
 
 let fallbackHeight = 30;
 
@@ -13,7 +14,11 @@ let fallbackHeight = 30;
 interface NumberControlProps {
   areNumberButtonsDisabled: boolean;
   updateEntry: (inputValue: number) => void;
-  getRemainingCellCountOfValue: (inputValue: number) => number;
+  getRemainingCellCountOfValue: (
+    sudokuBoard: SudokuObjectProps,
+    inputValue: number
+  ) => number;
+  sudokuBoard: SudokuObjectProps;
 }
 
 const NumberControl = (props: NumberControlProps) => {
@@ -21,6 +26,7 @@ const NumberControl = (props: NumberControlProps) => {
     areNumberButtonsDisabled,
     updateEntry,
     getRemainingCellCountOfValue,
+    sudokuBoard,
   } = props;
   const cellSize = getCellSize();
   const theme = useTheme();
@@ -59,8 +65,8 @@ const NumberControl = (props: NumberControlProps) => {
                   theme.colors.primaryContainer,
                 ]}
                 locations={[
-                  1 - getRemainingCellCountOfValue(number) / 9,
-                  1 - getRemainingCellCountOfValue(number) / 9,
+                  1 - getRemainingCellCountOfValue(sudokuBoard, number) / 9,
+                  1 - getRemainingCellCountOfValue(sudokuBoard, number) / 9,
                 ]}
                 style={{
                   width: cellSize
